+++ src/routes/[[lang]]/(protected)/credential-offer/+page.ts
@@ -2,26 +2,33 @@ import { debugPopupContent } from '$lib/components/organisms/debug/debug';
 import { credentialOfferStore } from '$lib/credentialOfferStore';
 import { m } from '$lib/i18n';
 import { callPar, holderQrToWellKnown, type QrToWellKnown } from '$lib/openId4vci';
+import { setCredentialAuthenticationPreference } from '$lib/preferences/credentialAuthentication';
 import type { Feedback } from '$lib/utils/types';
 import { get } from 'svelte/store';
 
 export const load = async () => {
 	const credentialOffer = get(credentialOfferStore);
 	let feedbackData: Feedback | undefined;
+
+	if (!credentialOffer) {
+		feedbackData = {
+			type: 'error',
+			feedback: 'no credential offer found'
+		};
+		return { feedbackData };
+	}
 	let wn: QrToWellKnown | undefined;
 	try {
 		wn = await holderQrToWellKnown(credentialOffer);
-	} catch {
+	} catch(e) {
 		feedbackData = {
 			type: 'error',
-			feedback: m.The_credential_issuer_is_currently_offline_you_may_try_again_later()
+			//@ts-ignore
+			message: e.message,
+			feedback: 'this service is not compatible or is currently offline'
 		};
 	}
 	if (!wn) {
-		feedbackData = {
-			type: 'error',
-			feedback: m.The_credential_issuer_is_currently_offline_you_may_try_again_later()
-		};
 		return { feedbackData };
 	}
 	const data = {
@@ -37,5 +44,11 @@ export const load = async () => {
 			feedback: m.The_credential_issuer_is_currently_offline_you_may_try_again_later()
 		};
 	}
+
+	await setCredentialAuthenticationPreference({
+		wn,
+		parResult
+	});
+
 	return { wn, authorizeUrl, parResult, feedbackData };
 };